refactor to redux:  guessWordForm
