+++ package.json
@@ -20,6 +20,7 @@
     "eslint-plugin-react": "6.4.1",
     "mocha": "^3.2.0",
     "react-addons-test-utils": "^15.4.2",
+    "redux-devtools": "^3.3.2",
     "sinon": "^2.1.0",
     "sinon-chai": "^2.9.0",
     "webpack": "^2.2.1",
@@ -29,7 +30,9 @@
     "babel": "^6.23.0",
     "plugin": "^0.3.3",
     "react": "^15.4.2",
-    "react-dom": "^15.4.2"
+    "react-dom": "^15.4.2",
+    "react-redux": "^5.0.3",
+    "redux": "^3.6.0"
   },
   "scripts": {
     "build": "webpack --config webpack.config.js",
