+++ src/index.js
@@ -1,7 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import {Provider} from 'react-redux';
+import {createStore} from 'redux';
+import reducer from './reducers';
 import App from './containers/App';
-ReactDOM.render(
-    <App />,
-    document.getElementById('root')
+
+let store = createStore(
+    reducer,
+    window.devToolsExtension ? window.devToolsExtension() : (f) => f
 );
+
+ReactDOM.render(
+    <Provider store={store}>
+        <App />
+    </Provider>,
+document.getElementById('root')
+)
+;
