+++ test/components/GuessNumberFromTest.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import GuessNumberForm from '../../src/containers/GuessNumberForm';
+import GuessNumberForm from '../../src/components/GuessNumberForm';
 describe('GuessNumberForm', () => {
     it('renders', () => {
         expect(shallow(
