+++ test/components/WordGameAppTest.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import WordGameApp from '../../src/components/WordGameApp';
-import GuessWordForm from '../../src/containers/GuessWordForm';
+import WordGameApp from '../../src/containers/WordGameApp';
+import GuessWordForm from '../../src/components/GuessWordForm';
 import WordHintList from '../../src/components/WordHintList';
 describe('WordGameApp', () => {
